Move UTP code to its own directory
