+++ test/link/test_llarp_link.cpp
@@ -1,8 +1,8 @@
-#include <link/utp.hpp>
 #include <link/iwp.hpp>
 #include <ev/ev.h>
 #include <messages/link_intro.hpp>
 #include <messages/discard.hpp>
+#include <utp/utp.hpp>
 
 #include <crypto/crypto_libsodium.hpp>
